refactor: Split code, create a separate, reusable, Ratings component
