+++ src/components/ListView/ListItem.js
@@ -1,34 +1,16 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { Link } from "react-router-dom";
 import { useSelector } from "react-redux";
-import { useSession } from "../../contexts/auth-context";
 import classes from "./ListItem.module.scss";
 
-import { Rating } from "@mui/material";
-import { rateProduct } from "../../helpers/rate-product";
+import Ratings from "../Ratings";
+
 import Controls from "../../components/Controls";
 
 const ListItem = (props) => {
-  const [readOnly, setReadOnly] = useState(false);
-  const [stars, setStars] = useState(0);
   const { product } = props;
-  const { user } = useSession();
-  const {
-    active,
-    arrivalDate,
-    brand,
-    category,
-    color,
-    description,
-    name,
-    price,
-    productCode,
-    productImage,
-    quantity,
-    rating,
-    reviews,
-    tags,
-  } = product;
+  const { color, description, name, price, productCode, productImage } =
+    product;
 
   const currency = useSelector((store) => store.productsReducer.currency);
 
@@ -41,28 +23,6 @@ const ListItem = (props) => {
       ></span>
     );
   });
-
-  const numberOfReviews = rating.ratings.length;
-  const averageRating =
-    rating.ratings.reduce((acc, cur) => acc + cur, 0) / numberOfReviews || 0;
-
-  useEffect(() => {
-    if (!user) setReadOnly(true);
-    if (user && rating.raters.includes(user.uid)) setReadOnly(true);
-    setStars(Math.round(averageRating / 10) / 2);
-  }, [user, rating, averageRating]);
-
-  const handleRating = (e) => {
-    const givenRating = +e.target.value * 20;
-    const userId = user.uid;
-    rateProduct(product, givenRating, userId);
-    setReadOnly(true);
-    setStars((prevStars) => {
-      const newAverage =
-        (averageRating * numberOfReviews + givenRating) / (numberOfReviews + 1);
-      return Math.round(newAverage / 10) / 2;
-    });
-  };
 
   return (
     <div className={classes.item}>
@@ -81,15 +41,7 @@ const ListItem = (props) => {
             2
           )}`}</span>
           <span className={classes.rating}>
-            <Rating
-              name="half-rating-red"
-              value={stars}
-              max={5}
-              precision={0.5}
-              size="small"
-              readOnly={readOnly}
-              onChange={handleRating}
-            />
+            <Ratings product={product} />
           </span>
         </div>
         <p>{description}</p>
