+++ src/components/Ratings/index.js
@@ -1,0 +1,2 @@
+import Ratings from "./Ratings";
+export default Ratings;
